step3: 리뷰 반영 - 객체 분리
