+++ README.md
@@ -3,7 +3,7 @@
 ## 구현한 기능 목록
 - 사용자의 입력을 받아 자동차 수와 이동을 시도할 횟수를 정의한다.
 - 자동차들에게 이동 명령을 내린다. 
-- 이동 명령을 받은 자동차들은 60%의 확률로 이동에 성공한다.
+- 전진하는 조건은 0에서 9 사이에서 random 값을 구한 후 random 값이 4이상일 경우이다.
 - 자동차들의 이동 기록을 모두 모아 데이터화 시킨 후에, 해당 데이터를 기반으로 모든 라운드의 자동차 위치를 출력한다.
 
 ## 진행 방법
