+++ src/main/java/com/kakao/racingcar/config/CarConfig.java
@@ -1,0 +1,10 @@
+package com.kakao.racingcar.config;
+
+public class CarConfig {
+    private CarConfig() {
+    }
+
+    public static final int MOVE_PERCENTAGE = 4;
+    public static final int BASE_PERCENTAGE = 10;
+    public static final int BEGIN_INDEX = 0;
+}
