+++ src/main/java/com/kakao/racingcar/domain/RacingGame.java
@@ -1,27 +1,33 @@
 package com.kakao.racingcar.domain;
 
 import com.kakao.racingcar.history.RacingHistory;
+import com.kakao.racingcar.util.RandomNumberGenerator;
+
 import java.util.stream.IntStream;
 
-import static com.kakao.racingcar.domain.CarConfig.BEGIN_INDEX;
-import static com.kakao.racingcar.domain.CarConfig.randomNumberGenerator;
+import static com.kakao.racingcar.config.CarConfig.BEGIN_INDEX;
 
 public class RacingGame {
     private final int tryCount;
     private final CarCollection carCollection;
+    private final RacingHistory racingHistory;
 
     public RacingGame(int tryCount, int totalCars) {
         this.tryCount = tryCount;
         this.carCollection = new CarCollection(totalCars);
+        this.racingHistory = new RacingHistory();
     }
 
     public void runRace() {
         IntStream.range(BEGIN_INDEX, tryCount).boxed()
-                .forEach(x -> carCollection.tryMoveCars(x, randomNumberGenerator(carCollection.size())));
+                .forEach(round -> {
+                    carCollection.tryMoveCars(RandomNumberGenerator.generate(carCollection.size()));
+                    racingHistory.logging(round, carCollection.getCars());
+                });
     }
 
     public RacingHistory getRacingHistory() {
-        return carCollection.getRacingHistory();
+        return racingHistory;
     }
 
 }
