+++ src/main/java/com/kakao/racingcar/util/RandomNumberGenerator.java
@@ -1,21 +1,21 @@
-package com.kakao.racingcar.domain;
+package com.kakao.racingcar.util;
 
 import java.util.List;
 import java.util.Random;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
 
-public class CarConfig {
-    private CarConfig() {
+import static com.kakao.racingcar.config.CarConfig.BASE_PERCENTAGE;
+import static com.kakao.racingcar.config.CarConfig.BEGIN_INDEX;
+
+public class RandomNumberGenerator {
+
+    private RandomNumberGenerator() {
     }
 
     private static final Random random = new Random();
 
-    public static final int MOVE_PERCENTAGE = 4;
-    public static final int BASE_PERCENTAGE = 10;
-    public static final int BEGIN_INDEX = 0;
-
-    public static List<Integer> randomNumberGenerator(int number) {
+    public static List<Integer> generate(int number) {
         return IntStream.range(BEGIN_INDEX, number).boxed()
                 .map(x -> random.nextInt(BASE_PERCENTAGE))
                 .collect(Collectors.toList());
