+++ src/test/java/com/kakao/racingcar/domain/CarCollectionTest.java
@@ -16,16 +16,4 @@ class CarCollectionTest {
         assertThat(carCollection.size()).isEqualTo(6);
     }
 
-    @Test
-    @DisplayName("입력한 숫자만큼의 차 저장소에서 이동을 완료한 후에, 차의 이동 기록을 볼 수 있다.")
-    void carCollectionRacingHistory() {
-        CarCollection carCollection = new CarCollection(3);
-        int tryMove = 10;
-        for (int i = 0; i < tryMove; i++) {
-            carCollection.tryMoveCars(i, Arrays.asList(10, 10, 0));
-        }
-        RacingHistory racingHistory = carCollection.getRacingHistory();
-        assertThat(racingHistory.getResult().size()).isEqualTo(tryMove);
-    }
-
 }
