+++ src/test/java/com/kakao/racingcar/domain/CarConfigTest.java
@@ -1,5 +1,7 @@
 package com.kakao.racingcar.domain;
 
+import com.kakao.racingcar.config.CarConfig;
+import com.kakao.racingcar.util.RandomNumberGenerator;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
 
@@ -12,14 +14,14 @@ class CarConfigTest {
     @Test
     @DisplayName("입력한 숫자 크기만큼의 랜덤 숫자를 생성한다.")
     void randomNumberGenerator() {
-        List<Integer> integers = CarConfig.randomNumberGenerator(10);
+        List<Integer> integers = RandomNumberGenerator.generate(10);
         assertThat(integers.size()).isEqualTo(10);
     }
 
     @Test
     @DisplayName("랜덥생성기에서 만들어진 값은 0이상 10미만이다")
     void randomNumberGeneratorValue() {
-        List<Integer> integers = CarConfig.randomNumberGenerator(100);
+        List<Integer> integers = RandomNumberGenerator.generate(100);
         for(int value : integers){
             assertThat(value >= 0 && value < 10).isTrue();
         }
